fix: Move VoteRecord into voting models
Addresses PR #1938 review comment #7.
